+++ app/roles/permissions.py
@@ -1,0 +1,10 @@
+from app.db.permissions import BasePermission
+from typing import List
+from app.business_partner.permissions import BusinessPartnerPermissions
+from app.users.permissions import UserPermissions
+
+
+AdminPermissions: List[BasePermission] = [
+    BusinessPartnerPermissions.list_()
+    + UserPermissions.list_()
+]
